# backup/20241230_000244/api-docs.yaml
openapi: 3.0.0
info:
  title: Heymarket API
  version: 1.0.0
  description: |
    API for interacting with Heymarket messaging service.
    Supports single messages, templates, and batch operations.

    ## Template Variables
    When using templates, variables are specified in {{variable}} format. Variables must match exactly between the template and the provided variables object.

    Example:
    ```
    Template: "Hello {{name}}, your points: {{points}}"
    Variables: { "name": "John", "points": "100" }
    ```

    Note: Dot notation (e.g. {{customer.name}}) requires the variable name to match exactly:
    ```
    Template: "Hello {{customer.name}}"
    Variables: { "customer.name": "John" }  # Must match template variable exactly
    ```
servers:
  - url: https://heymarket-api-dpbubhdceqb3fvge.canadaeast-01.azurewebsites.net
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/messages/range:
    get:
      summary: Get messages within a date range
      security:
        - ApiKeyAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Start date (ISO format)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: End date (ISO format)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
          description: Number of messages per page
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        phoneNumber:
                          type: string

    post:
      summary: Get messages within a date range with additional filtering
      security:
        - ApiKeyAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Start date (ISO format)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: End date (ISO format)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
          description: Number of messages per page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filters:
                  type: object
                  description: Additional filters for messages
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        phoneNumber:
                          type: string
  /api/messages/{phoneNumber}/all:
    get:
      summary: Get all messages for a specific phone number
      security:
        - ApiKeyAuth: []
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
          description: 10-digit phone number
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        status:
                          type: string
                        type:
                          type: string
                        channel:
                          type: string
  /api/messages/send:
    post:
      summary: Send a message to a phone number
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - message
              properties:
                phoneNumber:
                  type: string
                  description: Target phone number (10 or 11 digits). If 10 digits are provided, '1' will be prepended automatically.
                  example: "3476811000"
                message:
                  type: string
                  description: Message text to send
                  example: "Hello from the API!"
                attachments:
                  type: array
                  description: Optional array of media URLs
                  items:
                    type: string
                isPrivate:
                  type: boolean
                  description: Whether to create a private comment (memo) within the conversation
                  example: false
                author:
                  type: string
                  description: Display name for who the message was sent from
                  example: "Support Team"
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      messageId:
                        type: string
                      status:
                        type: string
                      timestamp:
                        type: string
                        format: date-time
                      to:
                        type: string
                      text:
                        type: string
  /api/messages/{phoneNumber}:
    get:
      summary: Get last message and customer info for a phone number
      security:
        - ApiKeyAuth: []
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
          description: 10-digit phone number
      responses:
        '200':
          description: Message info retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      customerId:
                        type: string
                      lastMessage:
                        type: object
                        properties:
                          id:
                            type: string
                          text:
                            type: string
                          date:
                            type: string
                            format: date-time
                          sender:
                            type: string
                          status:
                            type: string
                          type:
                            type: string
                          channel:
                            type: string
                      messageStatus:
                        type: string
                      totalMessages:
                        type: integer
                      phoneNumber:
                        type: string
...
  /api/batch/{batchId}:
    get:
      summary: Get batch operation status
      security:
        - ApiKeyAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                      progress:
                        type: object
                      timing:
                        type: object
                      errors:
                        type: object
                      metrics:
                        type: object
                        properties:
                          messages_per_second:
                            type: number
                          success_rate:
                            type: number
                          credits_used:
                            type: number
...